User controller

+++ src/main/java/org/koffa/beercrudjwtdb/services/BeerService.java
@@ -17,15 +17,13 @@ public class BeerService {
     public Beer findById(Long id) {
         return beerRepository.findById(id).orElseThrow(() -> new NoResultException("No beer found for id " + id));
     }
-    public void deleteById(Long id) {
-        beerRepository.deleteById(id);
+    public void deleteById(Long id) throws NoResultException {
+        if(beerRepository.existsById(id)) beerRepository.deleteById(id);
+        else throw new NoResultException("No beer found for id " + id);
     }
-    public Beer update(Long id, Beer beer) {
-        Beer beerToUpdate = findById(id);
-        beerToUpdate.setName(beer.getName());
-        beerToUpdate.setType(beer.getType());
-        beerToUpdate.setBrewery(beer.getBrewery());
-        return beerRepository.save(beerToUpdate);
+    public Beer update(Beer beer) {
+        if(!beerRepository.existsById(beer.getId())) throw new NoResultException("No beer found for id " + beer.getId());
+        return beerRepository.save(beer);
     }
     public Iterable<Beer> findAll() {
         return beerRepository.findAll();
